Extracted `HyperAdmin::ValidationHelper` module into its own file, which was overdue (my bad). Initially it contained a mere few lines of code, but in retrospect that’s not a valid excuse for breaking best practice...
